Add Indexer.Fetcher.Zkevm.BridgeL2
